Removing unused functions from utilities
- Moving viewport-dimensions into its own module

- Review fix: Remove outdated comment

- Review fix: removing cf_pagination, unused javascript in project
