+++ src/static/js/routes/common.js
@@ -23,7 +23,6 @@ $( document ).ready( function() {
   require( '../modules/reveal-on-focus' ).init();
   require( '../modules/scroll-on-history-collapse' ).init();
   require( '../modules/clear-form-buttons' ).init();
-  require( '../modules/jquery/cf_pagination' ).init();
   require( '../modules/jquery/custom-input' ).init();
   require( '../modules/jquery/custom-select' ).init();
   require( '../modules/footer-button' ).init();
